Add initial version of Exchange.
Also replaced Buy=1,Sell=2 with OrderType, updated BuyOrder,SellOrder
subclasses, and added a simple test.

+++ src/uk/ac/cam/teamOscarSSE/IDGenerator.java
@@ -1,11 +1,12 @@
 package uk.ac.cam.teamOscarSSE;
 
 public class IDGenerator {
-	static private int currentID = 1;
-	
-	private IDGenerator() {}
-	
-	public static int getID() {
+	static private long currentID = 1;
+
+	private IDGenerator() {
+	}
+
+	public static long getID() {
 		return currentID++;
 	}
 }
